0.2.6

+++ nonebot_plugin_who_at_me/config.py
@@ -7,4 +7,4 @@ class Config(BaseModel, extra=Extra.ignore):
     合并转发消息记录的超时时间, 单位为天
     """
 
-    reminder_expire_time: Optional[int] = None
+    reminder_expire_time: Optional[float] = None
